event: Setting meeting time is way too hard
Fix the way Gnome-Calendar interact with the user when the user is
about to create/edit an event.

Now the time widget will be restricted by the possibilities the user
has depending on the start time the user have setted to the event.

The time widget will try to adjust the range in between the start time
and end time based on the user requirement.

This requirement will be defined by the user as soon as the user sets a
time to finish the event. If the user needs to change the start time of this
event, the end time will follow the range the user have defined before.