Add --quit command line argument.
Overview: Clicking the close button in gnome-calendar doesn't always close all the instances of the application. The window is closed, but an instance of gnome-calendar is still running as a gapplication-service, which was probably run by gnome-shell search provider. To fix this, add --quit commandline argument that quits the running instance of the application.

The two changes made were as follows:

1) added the argument "quit" to GOptionEntry
2) added the definition of "quit" to gcal_application_command_line