gcal-date-chooser : Added bounds checking
Gnome calendar crashes when changing the date-seclector/date-chooser view from
a month with high number of day to one with lower number of days (example march
to february) when the selected day in the previous month is not found.

We can now introduce a bounds checking to ensure that the selected day number
in the previous month is found in the new month. And if it is not found we
should assigned it a closest number.

https://gitlab.gnome.org/GNOME/gnome-calendar/issues/212