month-view: Move event with CTRL key in DND
Currently there's not an easy way to copy an event from one day to
another.

This patch add a modifier key for the drag & drop event movement. With
this change the control key changes the default behaviour (move the
event) to copy so it's easier to copy an event in several days.

This solves the issue #34