src/gcal-event-widget.c tooltip dates/times changed to a human-readable format;
The format was changed to be more human-readable as per request.
Changes include include replacing month number with month name.
The format strings are now also surrounded in _(), because of special
formatting requirements of other languages. (e.g. czech dates: "22. ledna" require a "." after the day number.)

The currently used date/time format is not as readable.

Closes: https://gitlab.gnome.org/GNOME/gnome-calendar/issues/253