manager: fix reference management
GcalManager:remove_event() is an asynchronous function
that, by the time the callback is executed, unreferences
the event that is being removed. This implies that, when
the function was called, a reference was taken -- we don't
want the event to be finalized during the operation, after
all.

The problem is that the reference was never taken.

Fix that by properly owning a reference to the event when
removing it, and dropping it when the removal is finished.

Issue: #268