appmenu: Remove "New Event" entry
The "New Event" menu item is confusing and misleading, since
it is contextless item that has a context-aware action. The
result varies depending on which view you are, but this is in
no way communicated.

So just remove that, and all the GcalWindow API that is not
used anymore altogether.