Retire app-menu
For https://gitlab.gnome.org/GNOME/Initiatives/issues/4,
the application menu is being retired, with it's contents
being moved into the in-window primary window.

When removing a menu, the shortcuts for it's actions (if defined
in the menu) are also removed. The "Quit" shortcut is now
defined in src/gcal-window.c.

Related to https://gitlab.gnome.org/GNOME/gnome-calendar/issues/319