Stop using e_cal_util_get_system_timezone()
Because its return value is nullable, which is inconvenient for us.
Instead, add and use gcal_context_get_icaltimezone() to supplement the
existing gcal_context_get_timezone().

We'll also need a function to convert from GTimeZone to ICalTimezone,
gcal_timezone_to_icaltimezone().

Finally, we have to initialize the GcalSearchEngine later when
constructing GcalContext, because with this change GcalSearchEngine now
needs GcalContext ready to initialize itself.