weather-service: Enable met.no provider
As per the switch to libgweather 40 with commits 216a8984 and a2a6d2e4
we already comply with met.no TOS [1] as we already provide explicitly
the application contact info and (implicitly, being a GApplication) the
application-id.

So we can enable the met.no weather provider, making the weather
forecast to work again.

[1] https://api.met.no/doc/TermsOfService

Fixes: https://gitlab.gnome.org/GNOME/gnome-calendar/-/issues/682