manager: Fix bug where recurring events shift after being edited
Right now if a user edits say event 3 in a set of 5 events, the
whole event set gets shifted over by 3. It's as if the middle
instance event is treated as the new first event.

This is because of a bit of the code that clears the recurrence
id of the event, to get at the "main event", and edit the whole
set. The code retains the start time from the instance, though,
rather than use the start time of the main event it's trying to
update.

This commit queries for the main event from the instance event,
and sets the new main event start time to match the old main
event start time, before doing an update.