core: Handle exceptions in event list
If a recurrence of events is set up and one of the events gets
deleted, that event needs to be marked as an exception on the
main event so that it doesn't get recreated any time the main
event is edited.

This commit adds code to manage setting up this exception.