docs: Fix markdown commit message block in CONTRIBUTING
The code block for the commit message contains blank characters that are
breaking the markdown rendering engine, making the message block to
stop right after the first paragraph, and getting rid of the last line.

Simply cleaning these blank characters make the code block to render
properly. Note that I've also specified the type of the block, which is
plain text.