gcal-event-popover: Fix for wrapping title text in event popover summary
Previously, the event title header text in the details popover was being ellipsized(truncated) instead of wrapping, making it difficult to read longer titles.
This commit updates the gcal-event-popover.ui file to set the "wrap" and "wrap-mode" properties to "true" and "word-char", respectively, for the summary label object.
This ensures that the event title header text is wrapped and more readable in the summary popover.

Fixes #1016.