gcal-application: Ctrl+R to synchronize calendars
In !332, there was some difficulty with having two accelerators for the
"Synchronize calendars" action, so the Ctrl+R accelerator was left out.

This commit adds the Ctrl+R accelerator to gcal-window.ui by specifying
it before the F5 accelerator. The result is that only F5 is shown as an
accelerator in the overlay menu, but Ctrl+R still works.

Ctrl+R was also added to help-overlay.ui, displayed alongside F5.

Fixes #1052