gcal-calendars-page: Move new calendar button to bottom toolbar
Currently, if you have lots of calendars you need to scroll all the way
to the bottom of list to have access to the "New Calendar" button.

This commit moves the "New Calendar" button to be on a bottom toolbar
provided by AdwToolbarView. It makes the button always on view and easy
to access.

Closes https://gitlab.gnome.org/GNOME/gnome-calendar/-/issues/1051